danish_energy_project/requirements.txt
# Danish Energy Analytics Platform - Requirements

# Core data processing
pandas>=1.5.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Data transformation
dbt-core>=1.6.0
dbt-postgres>=1.6.0

# Machine learning
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
tensorflow>=2.13.0
joblib>=1.3.0

# Web framework and APIs
flask>=2.3.0
flask-cors>=4.0.0
flask-restx>=1.2.0
flask-limiter>=3.5.0
flask-socketio>=5.3.0

# Data visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities and helpers
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
click>=8.1.0
pyyaml>=6.0.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0

# Monitoring and logging
prometheus-client>=0.17.0
psutil>=5.9.0

# Time series and forecasting
statsmodels>=0.14.0
prophet>=1.1.0

# Caching and performance
redis>=4.6.0
gunicorn>=21.2.0

# Security
cryptography>=41.0.0
bcrypt>=4.0.0

